Extract LocalAuthorityClient from gateway_tests (#1492)
